Final Update before work on compression starts. changelog 1) Changed integer with cstdint for universal support 2) added android support 3) Some unneded logging is kept in createfile for developement.

+++ utils/bmp/reader/bmp_reader.h
@@ -4,6 +4,7 @@
 #include "../writer/bmp_writer.h"  // Include the BMP writer header file for the BMPFileHeader and BMPInfoHeader structures
 #include <fstream>
 #include <cmath>
+#include <cstdint>
 
-void readBMP(const std::string& filename, const std::string& outputFilename, unsigned long long int binary_length);
+void readBMP(const std::string& filename, const std::string& outputFilename, uint_fast64_t binary_length);
 #endif  // BMP_READER_H
